Standardized Readme
I added some badges, a maintainer section, a license section, a contribute section - with links to the contributing and CoC doc in multiformats/multiformats - and changed the empty github description.

I'm not quite sure what to put in install and usage; can you help there?

This is part of multiformats/multiformats#13.